requirements.txt
transformers==4.49.0
sentence-transformers==3.4.1
langchain==0.3.20
langchain-huggingface==0.1.2
langchain-community==0.3.19
huggingface-hub==0.29.3
transformers==4.49.0
streamlit==1.43.2
faiss-cpu==1.10.0
pypdf
python-dotenv
